requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
feedparser==6.0.11
redis==5.0.1
build==1.2.2.post1
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
cryptography==45.0.3
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.115.9
filelock==3.18.0
firebase-admin==6.8.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.5.1
# anthropic==0.39.0  # Removed - reverting to Gemini only
google-generativeai>=0.8.0  # Direct Gemini API instead of Vertex AI
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage>=1.32.0,<3.0.0
google-cloud-aiplatform>=1.38.0  # For Vertex AI image generation
google-cloud-secret-manager>=2.18.0  # For secure API key management
pillow>=11.2.0  # For image processing
uvicorn>=0.34.3  # For ASGI server (already included but ensuring version)
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.2
holidays==0.73
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.10.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
networkx==3.2.1
newsdataapi==0.1.23
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
#ollama==0.5.1
onnxruntime==1.19.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pillow==11.2.1
posthog==4.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
#tqdm==4.67.1
#transformers==4.52.4
#triton==3.3.0
tqdm==4.67.1 # Assuming tqdm and transformers are still needed, triton was the problem
transformers==4.52.4 # Assuming tqdm and transformers are still needed, triton was the problem
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.1
uritemplate==4.1.1
urljoin==1.0.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
firebase-admin
openai==1.61.0